LPS-105138 Move ColorPalette to common component
